+++ src/main/java/de/samply/generator/MeldungGenerator.java
@@ -1,18 +1,9 @@
 package de.samply.generator;
+
 import de.samply.model.DiagnoseTyp;
-import de.samply.model.HistologieTyp;
-import de.samply.model.OBDS;
-import de.samply.model.OBDS;
-import de.samply.model.OBDS.MengeMelder;
-import de.samply.model.OBDS.MengePatient.Patient;
-import de.samply.model.OBDS.MengePatient.Patient.MengeMeldung;
 import de.samply.model.OBDS.MengePatient.Patient.MengeMeldung.Meldung;
-import de.samply.model.OPTyp;
-import de.samply.model.STTyp;
 import de.samply.model.TumorzuordnungTyp;
-import de.samply.probabilities.ProbabilityType;
 import de.samply.random.RandomValuesGenerator;
-import java.util.Random;
 import javax.xml.datatype.DatatypeConfigurationException;
 import javax.xml.datatype.DatatypeFactory;
 import javax.xml.datatype.XMLGregorianCalendar;
@@ -45,7 +36,7 @@ public class MeldungGenerator {
     meldung.setDiagnose(this.createDiagnose());
     meldung.setST(this.stGenerator.createST(birthdate));
     meldung.setSYST(this.systGenerator.createSYST(birthdate));
-    meldung.getDiagnose().setHistologie(this.histologieGenerator.createHistologie());
+    meldung.getDiagnose().setHistologie(this.histologieGenerator.createHistologie(birthdate));
     meldung.setOP(this.opGenerator.createOP());
     this.count++;
     return meldung;
@@ -53,12 +44,12 @@ public class MeldungGenerator {
 
   private TumorzuordnungTyp createTumorzuordnung(){
     TumorzuordnungTyp tumorzuordnungTyp = new TumorzuordnungTyp();
-    tumorzuordnungTyp.setTumorID(""+this.count);
+    tumorzuordnungTyp.setTumorID("" + this.count);
     return tumorzuordnungTyp;
   }
 
   private DiagnoseTyp createDiagnose(){
-    DiagnoseTyp diagnose = new DiagnoseTyp();
-    return diagnose;
+    return new DiagnoseTyp();
   }
+  
 }