+++ src/main/java/de/samply/generator/SYSTGenerator.java
@@ -1,7 +1,6 @@
 package de.samply.generator;
 
 import de.samply.model.DatumTagOderMonatGenauTyp;
-import de.samply.model.DatumTagOderMonatOderJahrOderNichtGenauTyp;
 import de.samply.model.NebenwirkungTyp;
 import de.samply.model.NebenwirkungTyp.MengeNebenwirkung;
 import de.samply.model.SYSTTyp;
@@ -22,18 +21,19 @@ public class SYSTGenerator {
   private int count = 0;
   private final DatatypeFactory datatypeFactory = DatatypeFactory.newInstance();
   private final NebenwirkungGenerator nebenwirkungGenerator;
-  private MengeNebenwirkung nebenwirkungen;
+
   public SYSTGenerator(RandomValuesGenerator randomValuesGenerator) throws DatatypeConfigurationException {
     this.nebenwirkungGenerator = new NebenwirkungGenerator(randomValuesGenerator);
-    this.nebenwirkungen = new MengeNebenwirkung();
   }
 
   public SYSTTyp createSYST(XMLGregorianCalendar birthdate){
     SYSTTyp syst = new SYSTTyp();
     syst.setProtokoll(this.generateProtokoll());
     syst.setNebenwirkungen(this.generateNebewirkungen());
-    syst.setBeginn(generateBeginnDatum(birthdate));
-    syst.setEnde(syst.getBeginn().getValue());
+    if(birthdate.getYear()<2024){
+      syst.setBeginn(generateBeginnDatum(birthdate));
+      syst.setEnde(generateEndeDatum(syst.getBeginn().getValue()));
+    }
     return syst;
   }
   private String generateProtokoll() {
@@ -44,7 +44,8 @@ public class SYSTGenerator {
   private DatumTagOderMonatGenauTyp generateBeginnDatum(XMLGregorianCalendar birthdate) {
     DatumTagOderMonatGenauTyp beginnDatum = new DatumTagOderMonatGenauTyp();
     Random randomDays = new Random();
-    int daysToAdd = randomDays.nextInt(18262);
+    Date currentDate = new Date();
+    int daysToAdd = randomDays.nextInt(((currentDate.getYear()+1900) - birthdate.getYear()) * 365);
     XMLGregorianCalendar beginn;
     GregorianCalendar minGregorianCalendar = new GregorianCalendar();
     minGregorianCalendar.set(birthdate.getYear(), birthdate.getMonth(), birthdate.getDay());
@@ -57,8 +58,18 @@ public class SYSTGenerator {
     beginnDatum.setValue(beginn);
     return beginnDatum;
   }
-  private Date generateEndeDatum(XMLGregorianCalendar beginn){
-    Date endeDatum = new Date(beginn.getYear(),beginn.getMonth(),beginn.getDay());
+  private XMLGregorianCalendar generateEndeDatum(XMLGregorianCalendar beginn){
+    Random randomDays = new Random();
+    int daysToAdd = randomDays.nextInt(30);
+    GregorianCalendar minGregorianCalendar = new GregorianCalendar();
+    minGregorianCalendar.set(beginn.getYear(),beginn.getMonth(),beginn.getDay());
+    minGregorianCalendar.add(Calendar.DAY_OF_MONTH,daysToAdd);
+    XMLGregorianCalendar endeDatum = null;
+    try {
+      endeDatum = DatatypeFactory.newInstance().newXMLGregorianCalendar(minGregorianCalendar);
+    } catch (DatatypeConfigurationException e) {
+      throw new RuntimeException(e);
+    }
     return endeDatum;
   }